chore(polkit): change auth_admin_keep to auth_admin
When oma is run with pkexec to elevate privileges,
the PID of the oma process launched by pkexec is not consistent.
As a result, `auth_admin_keep` does not work as intended,
and users are still prompted to enter their password twice.

Since `auth_admin_keep` does not have the desired effect
in this scenario,
this commit replace it with `auth_admin`.

Signed-off-by: Chen Linxuan <me@black-desk.cn>