chore(data): split Debian (non-AOSC) 50oma.conf
Split 50oma-debian.conf from 50oma.conf so that TUM is not fetched. This
was not supposed to be an issue but with the following snippet of APT
configuration from Ubuntu 22.04:

  deb http://security.ubuntu.com/ubuntu/ jammy-security main restricted
  deb http://security.ubuntu.com/ubuntu/ jammy-security universe
  deb http://security.ubuntu.com/ubuntu/ jammy-security multiverse

When 50oma.conf is configured to fetch TUM data (updates.json),
`apt update' would throw the following warnings:

  W: Target TUM (updates.json) is configured multiple times in /etc/apt/sources.list:1 and /etc/apt/sources.list:2
  W: Target TUM (updates.json) is configured multiple times in /etc/apt/sources.list:1 and /etc/apt/sources.list:3

We will have to investigate this further but for now, split out the
configuration as a workaround.