fix(oma-console): use "reverse" ASCII sequence for search highlighting (#457)
Previously, we used "light gray" background for highlighting, which worked
well enough for coloured texts. But with `oma why', we started using light
gray texts without colouring and this made match highlighting unusable.
This could also have been problematic for light terminals.

Use the "reverse" sequence (7m) for highlighting so that the highlighting
is always readable, as we swap the foreground/background colours.