ci: fix check for oma-preview branch existence, rebase-push logic (#511)
The last implementation checked for a local `oma-preview' branch with
`git rev-parse', this is not correct as that branch was never checked out.
This caused the workflow to create a new `oma-preview' branch
unconditionally, destroying previous history in that process.

Also, rebased branches must be pushed forcefully, so specify `--force'
with the push command.